requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
sqlalchemy
pandas
python-dotenv
psycopg2-binary
openai
pytest
streamlit
plotly